Changed trials to store long

+++ app/src/main/java/xyz/kotlout/kotlout/model/adapter/TrialListAdapter.java
@@ -135,11 +135,11 @@ public class TrialListAdapter extends BaseExpandableListAdapter {
         break;
       case NON_NEGATIVE_INTEGER:
         NonNegativeTrial nonNegativeTrial = (NonNegativeTrial) trial;
-        trialResult.setText(Integer.toString(nonNegativeTrial.getResult()));
+        trialResult.setText(Long.toString(nonNegativeTrial.getResult()));
         break;
       case COUNT:
         CountTrial countTrial = (CountTrial) trial;
-        trialResult.setText(Integer.toString(countTrial.getResult()));
+        trialResult.setText(Long.toString(countTrial.getResult()));
         break;
       case MEASUREMENT:
         MeasurementTrial measurementTrial = (MeasurementTrial) trial;
